Move auth to new folder
